refactor(api-sub-agent): extract domains into modular structure
Extract 5 domain modules from api-sub-agent.js (875 LOC -> 210 LOC wrapper):
- endpoint-analysis.js: Endpoint detection and analysis
- documentation-analysis.js: OpenAPI/Swagger validation
- structure-analysis.js: RESTful patterns and versioning
- security-analysis.js: Security checks and CORS
- quality-analysis.js: Error handling, schemas, rate limiting

Changes:
- Create lib/agents/api-sub-agent/domains/ directory
- Extract domain-specific functions to separate modules
- Refactor main file to thin wrapper delegating to domains
- Add re-exports for backward compatibility
- Add 19 unit tests verifying modular structure

Part of SD-LEO-REFAC-API-AGENT-003
